requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
pyyaml>=6.0
Pillow>=9.0.0
numpy>=2.2.5
pip>=25.1.1
filelock>=3.18.0
pytz>=2025.2
torch>=2.7.0
ultralytics-thop>=2.0.14
requests>=2.32.3
charset-normalizer>=3.4.2
setuptools>=80.4.0
pyinstaller>=6.13.0
scipy>=1.15.3
matplotlib>=3.10.3
mpmath>=1.3.0
sympy>=1.14.0
tqdm>=4.67.1
torchvision>=0.22.0
psutil>=7.0.0
packaging>=25.0
networkx>=3.4.2
urllib3>=2.4.0
fsspec>=2025.3.2
MarkupSafe>=3.0.2
pandas>=2.2.3
python-dateutil>=2.9.0.post0
certifi>=2025.4.26
py-cpuinfo>=9.0.0
seaborn>=0.13.2
cycler>=0.12.1
altgraph>=0.17.4
six>=1.17.0
macholib>=1.16.3
idna>=3.10
contourpy>=1.3.2
fonttools>=4.58.0
pyparsing>=3.2.3
Jinja2>=3.1.6
ttkthemes>=3.2.2
kiwisolver>=1.4.8
pyinstaller-hooks-contrib>=2025.4
black>=25.1.0
platformdirs>=4.3.8
pathspec>=0.12.1
click>=8.2.0
isort>=6.0.1
mypy_extensions>=1.1.0
typing_extensions>=4.13.2